pyproject.toml
[project]
name = "full-auto-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.3.0",
    "APScheduler>=3.10.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.24.1",
    "persist-queue>=1.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
